DictionaryExt.  Other Ext files migrated to Common
