+++ Wabbajack.Common/Extensions/TaskExt.cs
